requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
boto3==1.34.49
botocore==1.34.49
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
fastapi-storages==0.3.0
fastapi-versioning==0.10.0
flower==2.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
Mako==1.3.1
MarkupSafe==2.1.4
orjson==3.9.12
packaging==24.0
passlib==1.7.4
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.27.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
